Client can read & display text from pipe + CLI opt
The client does non-blocking IO on stdin and on its pipe. It can read an
display text that is given on its pipe.
The client supports help `-h` and version `-v` options.
The server's pipe can be given as the client's first CLI argument, or
the client will use `/tmp/tchatche/server`.

When a server is launched, if the file `/tmp/tchatche/server` is
available, the server will symlink it to its pipe.